refactor(movex-core-util): :recycle: move out the unneeded fns
