+++ libs/movex-core-util/package.json
@@ -1,6 +1,6 @@
 {
   "name": "movex-core-util",
-  "version": "0.1.3-2",
+  "version": "0.1.3-5",
   "description": "Movex Core Util is the library of utilities for Movex",
   "license": "MIT",
   "author": {
