+++ libs/movex-core-util/src/lib/checkedState.ts
@@ -1,7 +1,5 @@
-import { MD5 } from 'object-hash';
-import { compare, applyReducer, deepClone } from 'fast-json-patch';
-import type { CheckedState, JsonPatch } from './core-types';
-import { isObject } from './misc';
+import { MD5 } from 'object-hash'; // TODO: If I can get a smaller version of this would be golded
+import type { CheckedState } from './core-types';
 
 export const computeCheckedState = <T>(state: T): CheckedState<T> => [
   state,
@@ -16,55 +14,3 @@ export const checkedStateEquals = <
   b: B
   // They are the same if the instances are the same [0] or if the checksums are the same [1]
 ) => a[0] === b[0] || a[1] === b[1];
-
-export const getJSONPatchDiff = <
-  A extends Record<string, any> | any[],
-  B extends Record<string, any> | any[]
->(
-  a: A,
-  b: B
-) => compare(a, b);
-
-export const applyMovexStatePatches = <TState>(
-  state: TState,
-  patchesInOrder: JsonPatch<TState>[]
-): TState => {
-  const allPatchesInOrder = patchesInOrder.reduce(
-    // Here there could be some logic on finding the same path in the diff,
-    //  and apply some optimization or heuristic for the reconciliation
-    (accum, next) => [...accum, ...next],
-    []
-  );
-
-  return allPatchesInOrder.reduce(
-    applyReducer,
-    // TODO: This is expensive but otherwise the state gets mutated. Need to look into maybe another way?
-    deepClone(state)
-  );
-};
-
-export const getMovexStatePatch = <A, B extends A>(
-  a: A,
-  b: B
-): JsonPatch<A> => {
-  if ((isObject(a) && isObject(b)) || (Array.isArray(a) && Array.isArray(b))) {
-    return getJSONPatchDiff(a, b);
-  }
-
-  // Primitives
-  // should this really be spporting primitives as well? Why not?
-  // TODO:
-  if (a !== b) {
-    return [
-      {
-        // Is this correct or what should happen in case of primitives?
-        path: '',
-        op: 'replace',
-        value: b,
-      },
-    ];
-  }
-
-  // TODO: Empty array if the same??
-  return [];
-};
